Add a content-length header to the response.
The length of the data is known at the time the response is returned,
and reporting it via a HTTP-ish header allows ftp responses to be more
easily used in code that also handles HTTP URLs.

The length is stored in an attribute on the data buffer and updated
during the ftplib data write callbacks.