Fix issues with Python 3.
Data must be converted to bytes before using socket.send. Text data
needs to be placed in a BytesIO buffer instead of StringIO and will be
converted back to str if needed by the Response object.  The type of the
auth data at any given point is just a complete disaster.